Create menu for simple algorithms

+++ calculator/SimpleAlgoritms.cs
@@ -16,18 +16,21 @@ namespace calculator
             }
             return sum == number ? true : false;
         }
-        public int PowerOfNumber(int number, int power)
+
+        public double PowerOfNumber(double number, int power)
         {
-            int result = 1;
+            double result = 1;
             for (int i = 0; i < power; i++) result *= number;
             return result;
         }
+
         public int FactorialNumber(int number)
         {
             int result = 1;
             for (int i = 1; i <= number; i++) result *= i;
             return result;
         }
+
         public bool PrimeNumber(int number)
         {
             bool isPrime = true;
@@ -44,13 +47,15 @@ namespace calculator
                 return true;
             return isPrime;
         }
-        public double AverageOfTheNumbers(int[] numbers)
+
+        public double AverageOfTheNumbers(double[] numbers)
         {
             double result = 0;
             for (int i = 0; i < numbers.Length; i++) result += numbers[i];
             return Math.Round(result / numbers.Length, 2);
         }
-        private int[] GetSortedArray(int[] numbers)
+
+        private double[] GetSortedArray(double[] numbers)
         {
             for (int i = 0; i < numbers.Length - 1; i++)
             {
@@ -58,7 +63,7 @@ namespace calculator
                 {
                     if (numbers[j] > numbers[j + 1])
                     {
-                        int tempValue = numbers[j];
+                        double tempValue = numbers[j];
                         numbers[j] = numbers[j + 1];
                         numbers[j + 1] = tempValue;
                     }
@@ -66,15 +71,18 @@ namespace calculator
             }
             return numbers;
         }
-        public int GetMinOfArray(int[] numbers)
+
+        public double GetMinOfArray(double[] numbers)
         {
             GetSortedArray(numbers);
             return numbers[0];
         }
-        public int GetMaxOfArray(int[] numbers)
+
+        public double GetMaxOfArray(double[] numbers)
         {
             GetSortedArray(numbers);
             return numbers[numbers.Length - 1];
         }
+
     }
 }
